add post.py
